refactor: 💡 form input submission
