+++ src/common.module.scss
@@ -1,5 +1,5 @@
-@value tblue #1d9bf0;
-@value tgray #868a8c;
+$tblue: #1d9bf0;
+$tgray: #868a8c;
 
 .input{
   border: solid 1px #cfd9de;
